Add files via upload
Initial commit; migration from internal code-base management server. 

NMR TOOLSUITE 0.1.0 (beta - fully working, but somewhat buggy)

Daq Muncher: A way of exploding a compact CSV into a series of files that represent NMR sweeps
Cronological Directory Sorter: A way of organizing the enormous number of files "Daq Muncher" creates
Sweep Averager: A way of averaging the enormous number of files in sub-directories created by "Cronological Directory Sorter", or manually sorted
NMR Analyzer: An signal extraction API for NMR data in the files created by the "Daq Muncher" or "Sweep Averager"
gui: A palatable front-end to "NMR Analyzer" so future researchers & undergrads don't need to put in the time on the command-line
global interpreter: A final program to collate the results of the NMR Analyzer into graphs that give a global impression of an experimental phase in the context of DNP enhanced targets.